# [SLM] Enable cuda graph in O3 mode