Cudagraph is enabled in O3 mode but the flag is not propagated to build config. This fixes it.

cc @Hzfengsy @tqchen